requirements.txt
# Dependências principais
osmnx==1.8.1
networkx==3.2.1
Pillow==10.2.0

# Dependências do osmnx (instaladas automaticamente)
# matplotlib==3.8.2
# numpy==1.26.3
# geopandas==0.14.1
# folium==0.15.1
# scipy==1.12.0